+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # Changelog
 
-## 3.2.1 (XXXX-XX-XX)
+## 3.2.1 (2024-02-02)
 ###### Release Highlights
 Changes to app UI as well as some quality of life changes. Added PIN code as a type for details.